+++ cryptopan/CHANGELOG.md
@@ -1,8 +1,8 @@
-<h1>Changelog</h1>
+# Changelog
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning]https://semver.org/spec/v2.0.0.html
 
-<h1>0.1.0-PRE (2022-01-23)</h1>
-<h2>Added</h2>
+# 0.1.0-PRE (2022-01-23)
+## Added
 Initial Version
