+++ cryptopan/Cargo.toml
@@ -10,9 +10,6 @@ categories = ["cryptography"]
 edition = "2021"
 readme = "README.md"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
 [dependencies]
 aes="0.8.1"
 bitvec = "1.0.1"
-ecb="0.1.1"
